BAU - refactor domain person identity package
